Lots of progress refactoring observing.
- Added class called `ObserveeSet` to hold all the models observed by a
given observer.
- Created a DeferredModelLoader which is a sort of proxy a model
classes (it whines unless you instantiate it with the stringy name of a
model) that also checks what `Object.const_defined?` has to say about
it, but it never actually tries to load the model unless explicitly
asked.
- Still need to set up inheritance and enablement and generally fix
ObserverArray which hasn't been ported for the updated code yet.